+++ domain/src/main/java/com/anytypeio/anytype/domain/subscriptions/GlobalSubscriptionManager.kt
@@ -1,6 +1,5 @@
 package com.anytypeio.anytype.domain.subscriptions
 
-import com.anytypeio.anytype.domain.chats.ChatPreviewContainer
 import com.anytypeio.anytype.domain.device.DeviceTokenStoringService
 import com.anytypeio.anytype.domain.device.NetworkConnectionStatus
 import com.anytypeio.anytype.domain.multiplayer.UserPermissionProvider
@@ -24,7 +23,6 @@ interface GlobalSubscriptionManager {
         private val profile: ProfileSubscriptionManager,
         private val networkConnectionStatus: NetworkConnectionStatus,
         private val deviceTokenStoringService: DeviceTokenStoringService,
-        private val chatPreviewContainer: ChatPreviewContainer,
         private val pushKeyProvider: PushKeyProvider
     ) : GlobalSubscriptionManager {
